==> Project and competition/Questionnaire/1_Data preprocess.ipynb <==
資料前處理

讀入套件及檔案

import pandas as pd
import numpy as np
import warnings

warnings.filterwarnings('ignore')

filename = 'D:/研究所/1132/智慧行銷管理/期末報告/Data.csv'
data = pd.read_csv(filename)

資料前處理(刪除不用欄位、改變欄位名稱)

cols_to_drop = ['時間戳記', '(呈上題) 針對選到 "會購買" 者\n請問您願意花費多少錢購買此產品?']
data = data.drop(cols_to_drop, axis = 1)

col_names = {'請問您的性別是：' : 'gender', 
             '您的年齡：' : 'age',
             '教育程度（學歷） ' : 'education', 
             ' 請問您的職業是:' : 'career', 
             '居住地區：' : 'living_location',
             '請問您每月平均所得為?' : 'monthly_income',
             '請問您每月平均支出為? ' : 'monthly_spending', 
             '請問您曾經使用過防曬商品嗎?' : 'used_product',
             '(呈上題) 針對選到 "一直有使用" 或 "曾經使用過" 者\n請問您使用防曬商品多久時間了?' : 'time_used',
             '(呈上題) 針對選到 "一直有使用" 或 "曾經使用過" 者\n請問您使用防曬商品的頻率?' : 'usage_frequency',
             '請問您最常使用哪些防曬商品?' : 'most_frequently_used_product',
             '請問您使用過哪些ANESSA 安耐曬商品?' : 'anessa_products_used',
             '購買意願 [1.\t未來我有意願購買ANESSA 安耐曬商品。]' : 'purchase_intention_q1',
             '購買意願 [2.\t我會考慮購買ANESSA 安耐曬商品是。]' : 'purchase_intention_q2',
             '購買意願 [3.\t我願意向他人推薦購買ANESSA 安耐曬商品]' : 'purchase_intention_q3',
             '品質價值 [1.\tANESSA 安耐曬商品具有一致的品質。]' : 'quality_value_q1',
             '品質價值 [2.\tANESSA 安耐曬商品製作過程良好。]' : 'quality_value_q2',
             '品質價值 [3.\tANESSA 安耐曬商品品質達到可接受的標準。]' : 'quality_value_q3',
             '品質價值 [4.\tANESSA 安耐曬商品品質良好。]' : 'quality_value_q4',
             '品質價值 [5.\tANESSA 安耐曬商品持久耐用。]' : 'quality_value_q5',
             '品質價值 [6.\tANESSA 安耐曬商品產品品質穩定]' : 'quality_value_q6',
             '情緒價值 [1.\tANESSA 安耐曬商品讓我感到享受的產品。]' : 'emotional_value_q1',
             '情緒價值 [2.\tANESSA 安耐曬商品會讓我產生想要使用的感覺。]' : 'emotional_value_q2',
             '情緒價值 [3.\t使用ANESSA 安耐曬商品會讓我感到放鬆。]' : 'emotional_value_q3',
             '情緒價值 [4.\tANESSA 安耐曬商品會讓我感覺良好。]' : 'emotional_value_q4',
             '情緒價值 [5.\tANESSA 安耐曬商品會帶給我愉悅的感受。]' : 'emotional_value_q5',
             '價格價值 [1.\tANESSA 安耐曬商品的價格是合理的。]' : 'price_value_q1',
             '價格價值 [2.\tANESSA 安耐曬商品具有物有所值的特性。]' : 'price_value_q2',
             '價格價值 [3.\t以這個價格來說，ANESSA 安耐曬商品是不錯的產品。]' : 'price_value_q3',
             '價格價值 [4.\tANESSA 安耐曬商品具有CP值。]' : 'price_value_q4',
             '社會價值 [1.\tANESSA 安耐曬商品有助於讓我感到被他人接納。]' : 'social_value_q1',
             '社會價值 [2.\tANESSA 安耐曬商品能改善他人對我的看法。]' : 'social_value_q2',
             '社會價值 [3.\tANESSA 安耐曬商品能使我在人際互動中留下良好印象。]' : 'social_value_q3',
             '社會價值 [4.\t擁有ANESSA 安耐曬商品能讓人獲得社會上的認可。]' : 'social_value_q4',
             '網路口碑 [1.\t對我而言，消費者網路上對ANESSA 安耐曬商品評論非常重要。]' : 'online_reviews_q1',
             '網路口碑 [2.\t我會特別搜尋網路上ANESSA 安耐曬商品 的相關資訊。]' : 'online_reviews_q2',
             '網路口碑 [3.\t網路上對ANESSA 安耐曬商品評論會對我的購買決策有關鍵性的影響。]' : 'online_reviews_q3',
             '網路口碑 [4.\t網路上對ANESSA 安耐曬商品評論，有助於我更準確地判斷產品的品質。]' : 'online_reviews_q4',
             '網路口碑 [5.\t網路上對ANESSA 安耐曬商品評論，在我的決策過程中扮演重要角色。]' : 'online_reviews_q5',
             '整體而言，您是否會購買ANESSA 安耐曬金鑽高效防曬露  ？ ' : 'overall_purchase_intention_anessa'}

data = data.rename(columns = col_names)

如果「請問您曾經使用過防曬商品嗎?」這一題不是「沒使用過」，在使用多久時間及頻率兩題不能為空值，如果是空值，則移除此筆資料

condition = data['used_product'].str.contains('沒有使用過', na = False)
condition_B = data['time_used'].isna() | data['usage_frequency'].isna()
data = data[~(~condition & condition_B)]

如果「請問您曾經使用過防曬商品嗎?」這一題為「沒使用過」，在使用多久時間及頻率兩題改為空值

data.loc[condition, ['time_used', 'usage_frequency']] = np.nan

資料整理特徵合併及編碼

gender_map = {'女' : 0,
              '男' : 1}
data['gender'] = data['gender'].map(gender_map)

age_map = {'18歲以下' : 0,
           '18-25歲' : 0,
           '26-30歲' : 1,
           '31-35歲' : 1,
           '36-40歲' : 2,
           '41-45歲' : 2,
           '46-50歲' : 3,
           '51-55歲' : 3,
           '61-65歲' : 3}
data['age'] = data['age'].map(age_map)

edu_map = {'國中以下' : 0,
           '高中(職)' : 0,
           '大學/大專' : 1,
           '四技五專' : 1,
           '碩士' : 2,
           '博士' : 2}
data['education'] = data['education'].map(edu_map)

career_map = {'學生' : 0,
              '專業人員(醫生、律師、工程師等)' : 1,
              '醫療' : 1,
              '資訊科技業' : 1,
              '軍公教人員(軍人、公務人員、教師)' : 1,
              '教育' : 1,
              '製造業' : 1,
              '金融業' : 1,                    
              '營造業' : 1,              
              '通訊業' : 1,          
              '電信' : 1,
              '零售業' : 1,             
              '傳播/文化業' : 1,
              '行政' : 1,
              '服務業' : 2,
              '人力仲介代徵' : 2,
              '自由業' : 2,
              '家庭主婦' : 2}
data['career'] = data['career'].map(career_map)

location_map = {'北部（台北、新北、基隆、桃園、新竹、苗栗）' : 0,
                '南部（嘉義、台南、高雄、屏東）' : 1,
                '中部（台中、彰化、南投、雲林）' : 1,
                '東部（宜蘭、花蓮、台東）' : 1,
                '離島地區(澎湖、金門、馬祖)' : 1}
data['living_location'] = data['living_location'].map(location_map)

income_map = {'$30,000元(含)以下' : 0,
              '$30,001-50,000元' : 1,
              '$50,001-70,000元' : 1,
              '$70,001-90,000元' : 1,
              '$90,001-110,000元' : 1,
              '$110,001-130,000元' : 1,
              '$130,001-150,000元' : 1,
              '$150,000元以上' : 1}
data['monthly_income'] = data['monthly_income'].map(income_map)

spending_map = {'$20,000元(含)以下' : 0,
                '$20,001-40,000元' : 1,
                '$40,001-60,000元' : 1,
                '$60,001-80,000元' : 1,
                '$80,001-100,000元' : 1,
                '$100,001-120,000元' : 1,
                '$120,000元以上' : 1}
data['monthly_spending'] = data['monthly_spending'].map(spending_map)

used_product_map = {'沒有使用過' : 0,
                    '沒有使用過，未來也不會使用' : 0,
                    '沒有使用過，但未來想要使用(不排斥)' : 0,
                    '曾經使用過' : 1,
                    '一直使用' : 2,
                    '一直有使用' : 2}
data['used_product'] = data['used_product'].map(used_product_map)

time_used_map = {'3個月以下' : 1,
                 '3-6個月' : 1,
                 '6-12個月' : 1,
                 '1-2 年' : 2,
                 '2-3年' : 2,
                 '3-4年' : 2,
                 '5-6年' : 2,
                 '6年以上' : 3}
data['time_used'] = data['time_used'].map(time_used_map).fillna(0)

usage_frequency_map = {'每天' : 1,
                       '兩到三天一次' : 2,
                       '四到五天一次' : 2,
                       '五到七天一次' : 2,
                       '一週一次' : 2,
                       '兩到三週一次' : 3,
                       '一個月一次' : 3,
                       '兩到三個月一次' : 3,
                       '四到六個月一次' : 3,
                       '七個月以上一次' : 3}
data['usage_frequency'] = data['usage_frequency'].map(usage_frequency_map).fillna(0)

condition = data['most_frequently_used_product'].str.contains('沒有使用', na = False)
data.loc[condition, ['most_frequently_used_product']] = np.nan
most_frequently_used_product_map = {'防曬乳液 (Sunscreen Lotion)' : 1,
                                    '防曬霜 (Sunscreen Cream)' : 1,
                                    '防曬噴霧 (Sunscreen Spray)' : 1,
                                    '防曬凝膠 (Sunscreen Gel)' : 1,
                                    '防曬乳液 (Sunscreen Lotion), 防曬噴霧 (Sunscreen Spray)' : 2,
                                    '防曬乳液 (Sunscreen Lotion), 防曬霜 (Sunscreen Cream)' : 2,
                                    '防曬霜 (Sunscreen Cream), 防曬噴霧 (Sunscreen Spray)' : 2,
                                    '防曬乳液 (Sunscreen Lotion), 防曬棒 (Sunscreen Stick)' : 2,
                                    '防曬噴霧 (Sunscreen Spray), 防曬凝膠 (Sunscreen Gel)' : 2, 
                                    '防曬乳液 (Sunscreen Lotion), 防曬霜 (Sunscreen Cream), 防曬噴霧 (Sunscreen Spray)' : 2,
                                    '防曬乳液 (Sunscreen Lotion), 防曬噴霧 (Sunscreen Spray), 防曬粉 (Sunscreen Powder)' : 2,
                                    '防曬乳液 (Sunscreen Lotion), 防曬霜 (Sunscreen Cream), 防曬棒 (Sunscreen Stick)' : 2,
                                    '防曬乳液 (Sunscreen Lotion), 防曬凝膠 (Sunscreen Gel), 防曬粉 (Sunscreen Powder)' : 2,
                                    '防曬乳液 (Sunscreen Lotion), 防曬霜 (Sunscreen Cream), 防曬噴霧 (Sunscreen Spray), 防曬粉 (Sunscreen Powder)' : 2}
data['most_frequently_used_product'] = data['most_frequently_used_product'].map(most_frequently_used_product_map).fillna(0)

condition = data['anessa_products_used'].str.contains('沒有用', na = False)
data.loc[condition, ['anessa_products_used']] = np.nan
anessa_products_used_map = {'水嫩寶貝敏感肌防曬凝膠' : 1,
                            '金鑽高效防曬露NA（5X版）' : 1,
                            '輕親柔光敏感肌防曬露NA': 1,
                            '曬後淨白修復精華（夜用）' : 1,
                            '金鑽高效水感防曬凝膠NA' : 1, 
                            '美光燈潤色防曬凝膠N' : 1,
                            '金鑽高效防曬露NA（5X版）, 輕親柔光敏感肌防曬露NA, 美光燈潤色防曬凝膠N' : 2,
                            '金鑽高效防曬露NA（5X版）, 金鑽高效水感防曬凝膠NA, 輕親柔光敏感肌防曬露NA' : 2,
                            '金鑽高效防曬露NA（5X版）, 輕親柔光敏感肌防曬露NA' : 2,
                            '金鑽高效水感防曬凝膠NA, 輕親柔光敏感肌防曬露NA' : 2,
                            '金鑽高效防曬露NA（5X版）, 金鑽高效水感防曬凝膠NA' : 2,
                            '金鑽高效防曬露NA（5X版）, 金鑽高效水感防曬凝膠NA, 曬後淨白修復精華（夜用）' : 2,
                            '金鑽高效防曬露NA（5X版）, 美光燈潤色防曬凝膠N' : 2,
                            '輕親柔光敏感肌防曬露NA, 美光燈潤色防曬凝膠N' : 2,
                            '金鑽高效水感防曬凝膠NA, 輕親柔光敏感肌防曬露NA, 美光燈潤色防曬凝膠N' : 2}
data['anessa_products_used'] = data['anessa_products_used'].map(anessa_products_used_map).fillna(0)

cols = ['purchase_intention_q1',
       'purchase_intention_q2', 'purchase_intention_q3', 'quality_value_q1',
       'quality_value_q2', 'quality_value_q3', 'quality_value_q4',
       'quality_value_q5', 'quality_value_q6', 'emotional_value_q1',
       'emotional_value_q2', 'emotional_value_q3', 'emotional_value_q4',
       'emotional_value_q5', 'price_value_q1', 'price_value_q2',
       'price_value_q3', 'price_value_q4', 'social_value_q1',
       'social_value_q2', 'social_value_q3', 'social_value_q4',
       'online_reviews_q1', 'online_reviews_q2', 'online_reviews_q3',
       'online_reviews_q4', 'online_reviews_q5']
scale_5_map = {'非常不同意' : 0,
               '不同意' : 1,
               '普通' : 2,
               '同意' : 3,
               '非常同意' : 4}
for col in cols:
    data[col] = data[col].map(scale_5_map)

overall_map = {'不會' : 0,
               '會' : 1}
data['overall_purchase_intention_anessa'] = data['overall_purchase_intention_anessa'].map(overall_map)

data.to_csv('D:/研究所/1132/智慧行銷管理/期末報告/Preprocessed_data.csv', index = False)

data
